requirements.txt
flask
tensorflow-cpu
gunicorn
pillow
numpy
opencv-python-headless
object_detection
tensorflow_hub